@@ -1,16 +1,16 @@
 {
   "name": "pixels-catcher",
   "version": "0.11.0",
-  "description": "UI snpashot testing for React Native",
+  "description": "UI snapshot testing for React Native",
   "main": "lib/client/index.js",
   "scripts": {
     "demo": "cd demo && ../node_modules/.bin/flow-node ../cli.js android debug",
-    "flow": "flow",
-    "lint": "eslint src",
-    "prepublish": "./mk_lib.sh",
-    "test": "jest",
-    "test-coverage": "jest --coverage",
-    "test-coverage-report": "codecov"
+    "lint": "eslint --ext .js --ext .ts --ext .tsx ./src/",
+    "build": "./node_modules/.bin/tsc -p src/client && ./node_modules/.bin/tsc -p src/runner",
+    "postbuild": "npm run assets",
+    "prepublish": "npm run build",
+    "prepack": "npm run build",
+    "assets": "cp ./src/runner/server/dummy.png ./lib/runner/server/"
   },
   "bin": "./lib/runner/cli.js",
   "files": [
@@ -24,42 +24,7 @@
   "author": "Maksym Rusynyk <m.rusynyk@gmail.com>",
   "license": "MIT",
   "engines": {
-    "node": ">=12.*"
-  },
-  "dependencies": {
-    "formidable": "^1.2.2",
-    "multer": "^1.4.2",
-    "pixelmatch": "^5.2.0",
-    "pngjs": "^5.0.0",
-    "react-native-save-view": "^0.2.3"
-  },
-  "peerDependencies": {
-    "react": ">=16.8.6",
-    "react-native": ">=0.60",
-    "prop-types": ">=15.7.2"
-  },
-  "devDependencies": {
-    "react": "16.8.6",
-    "react-native": "0.60.5",
-    "@babel/core": "^7.5.5",
-    "@babel/preset-flow": "^7.0.0",
-    "@babel/runtime": "^7.5.5",
-    "@react-native-community/eslint-config": "^0.0.5",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.9.0",
-    "eslint": "^6.2.0",
-    "jest": "^24.9.0",
-    "metro-react-native-babel-preset": "^0.56.0",
-    "react-test-renderer": "16.8.6",
-    "codecov": "^3.0.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-plugin-flowtype": "^3.2.1",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "^6.1.2",
-    "eslint-plugin-react": "^7.11.1",
-    "flow-bin": "0.98.0",
-    "flow-remove-types": "^1.2.3",
-    "flow-typed": "^2.5.1"
+    "node": ">=14.*"
   },
   "repository": {
     "type": "git",
@@ -68,5 +33,29 @@
   "bugs": {
     "url": "https://github.com/rumax/react-native-PixelsCatcher/issues"
   },
-  "homepage": "https://github.com/rumax/react-native-PixelsCatcher#readme"
+  "homepage": "https://github.com/rumax/react-native-PixelsCatcher#readme",
+  "dependencies": {
+    "formidable": "^1.2.2",
+    "pixelmatch": "^5.2.1",
+    "pngjs": "^4.0.1",
+    "react-native-save-view": "^0.2.3"
+  },
+  "devDependencies": {
+    "@types/formidable": "^1.2.2",
+    "@types/pixelmatch": "^5.2.3",
+    "@types/pngjs": "^6.0.0",
+    "@types/react-native": "^0.64.6",
+    "@typescript-eslint/eslint-plugin": "^4.25.0",
+    "@typescript-eslint/parser": "^4.25.0",
+    "eslint": "^7.27.0",
+    "eslint-config-airbnb": "^18.2.1",
+    "eslint-plugin-import": "^2.23.3",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-react": "^7.23.2",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "react": "17.0.1",
+    "react-native": "0.64.0",
+    "react-test-renderer": "17.0.1",
+    "typescript": "^4.2.4"
+  }
 }
