@@ -4,13 +4,10 @@
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 */
-/* @flow */
-const exec = require('./exec');
-
-function isCommand(cmd: string) {
-  const out = exec(`whereis ${cmd}`);
-
-  return Boolean(out.trim());
+function delay(milliseconds: number): Promise<unknown> {
+  return new Promise((resolve: Function) => {
+    setTimeout(resolve, milliseconds);
+  });
 }
 
-module.exports = isCommand;
+export default delay;
