@@ -4,12 +4,10 @@
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 */
-/* @flow */
-const exec = require('../exec');
-const isCommand = require('../isCommand');
+import exec from '../exec';
+import isCommand from '../isCommand';
 
-module.exports =
-  process.env.ANDROID_EMULATOR
+export default process.env.ANDROID_EMULATOR
   || (isCommand('emulator') ? 'emulator' : undefined)
   || (
     exec('uname -s').trim() === 'Darwin'
