@@ -4,11 +4,10 @@
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 */
-/* @flow */
-function delay(milliseconds: number): Promise<*> {
-  return new Promise((resolve: Function) => {
-    setTimeout(resolve, milliseconds);
-  });
-}
 
-module.exports = delay;
+const timeToSec = (ms: number): number => {
+  const sec = ms / 1000;
+  return Math.round(sec * 1000) / 1000;
+};
+
+export default timeToSec;
