@@ -1,11 +1,12 @@
 /**
-* Copyright (c) Maksym Rusynyk 2018 - present
-*
-* This source code is licensed under the MIT license found in the
-* LICENSE file in the root directory of this source tree.
-*/
-/* @flow */
+ * Copyright (c) Maksym Rusynyk 2018 - present
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+// eslint-disable-next-line no-use-before-define
 import React, { Component } from 'react';
+// @ts-ignore
 import SaveView from 'react-native-save-view';
 
 import { getNextSnapshot } from './snapshotsManager';
@@ -15,11 +16,11 @@ import network from './utils/network';
 
 const TAG = 'PIXELS_CATCHER::APP::SNAPSHOTS_CONTAINER';
 
-export default class SnapshotsContainer extends Component<*, *> {
+export default class SnapshotsContainer extends Component<any, any> {
   // eslint-disable-next-line react/sort-comp
-  _viewRef: ?React$ElementRef<*>;
+  _viewRef: any;
 
-  _testStartedAt: number = (new Date()).getTime();
+  _testStartedAt: number = new Date().getTime();
 
   _renderStartedAt: number = 0;
 
@@ -40,7 +41,6 @@ export default class SnapshotsContainer extends Component<*, *> {
     this._onRef = this._onRef.bind(this);
   }
 
-
   render() {
     const { ActiveSnapshot } = this.state;
 
@@ -51,21 +51,17 @@ export default class SnapshotsContainer extends Component<*, *> {
 
     log.i(TAG, `rendering snapshot [${ActiveSnapshot.snapshotName}]`);
 
-    this._renderStartedAt = (new Date()).getTime();
+    this._renderStartedAt = new Date().getTime();
 
-    return (
-      <ActiveSnapshot ref={this._onRef} onReady={this._onSnapshotReady} />
-    );
+    return <ActiveSnapshot ref={this._onRef} onReady={this._onSnapshotReady} />;
   }
 
-
-  _onRef(ref: ?React$ElementRef<*>) {
+  _onRef(ref: any) {
     this._viewRef = ref;
   }
 
-
   _onSnapshotReady() {
-    const renderTime = (new Date()).getTime() - this._renderStartedAt;
+    const renderTime = new Date().getTime() - this._renderStartedAt;
     log.v(TAG, 'Snapshot ready');
 
     setTimeout(async () => {
@@ -96,7 +92,7 @@ export default class SnapshotsContainer extends Component<*, *> {
         return;
       }
 
-      let failure;
+      let failure: any;
 
       try {
         log.v(TAG, '++SaveView.save');
@@ -125,11 +121,9 @@ export default class SnapshotsContainer extends Component<*, *> {
     }, 50);
   }
 
-
   _getTestExecutionTime(): number {
-    return (new Date()).getTime() - this._testStartedAt;
+    return new Date().getTime() - this._testStartedAt;
   }
-
 
   nextSnapshot() {
     log.v(TAG, 'Trying to gen next snapshot');
@@ -137,7 +131,7 @@ export default class SnapshotsContainer extends Component<*, *> {
 
     if (nextSnapshot) {
       log.v('Switching to next snapshot');
-      this._testStartedAt = (new Date()).getTime();
+      this._testStartedAt = new Date().getTime();
       this.setState({ ActiveSnapshot: nextSnapshot });
     } else {
       log.v('No more snapshots left, exit testing');
