@@ -4,10 +4,9 @@
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 */
-/* @flow */
 import log from './utils/log';
 
-const snapshots = [];
+const snapshots: any = [];
 const TAG = 'PIXELS_CATCHER::APP::SNAPSHOTS_MANAGER';
 
 
