@@ -4,11 +4,10 @@
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 */
-/* @flow */
 import network from './network';
 
 const compareToReference = async (snapshotName: string, base64: string): Promise<string | void> => {
-  const response = await network.postBase64({
+  const response: any = await network.postBase64({
     base64,
     fileName: `${snapshotName}.png`,
   });
