@@ -1,15 +1,15 @@
+/* eslint-disable no-unused-vars */
 /**
 * Copyright (c) Maksym Rusynyk 2018 - present
 *
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 */
-/* @flow */
 
 export type StartParamsType = Array<string>;
 
 export interface DeviceInterface {
-  constructor(deviceName: string, canStopDevice?: boolean): void,
+  // constructor(deviceName: string, canStopDevice?: boolean): void,
 
   start(params: StartParamsType): Promise<void>,
 
